% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_info.R
\name{get_network_info}
\alias{get_network_info}
\title{Get network info from Web of Life}
\usage{
get_network_info(type, names = "yes")
}
\arguments{
\item{type}{interaction type}

\item{names}{would you like to include the names of species?}
}
\value{
list of networks
}
\description{
Get network info from Web of Life
}
