% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_info.R
\name{get_network_info}
\alias{get_network_info}
\title{Get network info from Web of Life}
\usage{
get_network_info(interaction_type, data_type)
}
\arguments{
\item{interaction_type}{"all", "pollination", "seed disperser", "plant-ant", "plant-herbivore", "anemone-fish", "host-parasite", "food web"}

\item{data_type}{"binary", "weighted", or "all"}
}
\value{
data frame with metadata for each network.
}
\description{
Get network info from Web of Life
}
