% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{get_networks}
\alias{get_networks}
\title{Get networks from Web of Life}
\usage{
get_networks(network_name, interaction_type, data_type, names)
}
\arguments{
\item{network_name}{Character}
\item{interaction_type}{plant-ant", "pollination", "seed disperser", "plant-herbivore", "anemone-fish", "host-parasite", "food web", "all" (default)}

\item{data_type}{"binary", "weighted", or "all" (default)}

\item{names}{would you like to include the names of species? default = "yes", otherwise input "no"}
}
\value{
list of networks
}
\description{
Get networks from Web of Life
}
