% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{get_networks}
\alias{get_networks}
\title{Get networks from Web of Life}
\usage{
get_networks(interaction_type, data_type, names = "yes")
}
\arguments{
\item{interaction_type}{"all", "plant-ant", "pollination", "seed disperser", "plant-herbivore", "anemone-fish", "host-parasite", "food web"}

\item{data_type}{"binary", "weighted", or "all"}

\item{names}{would you like to include the names of species? default = "yes", otherwise input "no"}
}
\value{
list of networks
}
\description{
Get networks from Web of Life
}
