% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{get_networks}
\alias{get_networks}
\title{Get networks from Web of Life}
\usage{
get_networks(type, names = "yes")
}
\arguments{
\item{type}{interaction type}

\item{names}{would you like to include the names of species?}
}
\value{
list of networks
}
\description{
Get networks from Web of Life
}
